requirements.txt
python-dotenv==1.0.0
pytz==2023.3
telethon==1.34.0
apscheduler==3.10.4
supabase==2.8.0
httpx==0.27.0